Changed public key to authorized key format (December 5, 2020)
This release changes the `publicKeyBase64` field to `publicKey` and the format from the OpenSSH wire format to the authorized_keys format.

This also means that the internal API format changes from `[]byte` to `string`.

Transitioning to the authorized_keys format should make it easier for auth server implementers to authenticate against SSH keys.